Return all prompts,add  random keyword any
